Ensure that `USER` is set when building packages.
We've run across a build which fails if USER is not set in the
environment, `astyle`, so we'll explicitly set it for consistency with
what a normal interactive session would have.

I've diff `env` in the container environment and one in a normal
interactive oneother and this is the only difference.